See description
C++:

- Lots of QOL improvements: Moved SDL out of the Xcode project, and we’re now using the framework from the frameworks folder (results in faster build times)
- Timeline, TimeTrack for animating objects across a timeline using keyframes
- `AnimatedSpriteRenderer` improvements
- New: `CharacterController`, `BehaviorNode`, for building complex character behaviors
- New: `SimpleAnimationController` for linking states to frame animations (2D)
- New: `ComponentsController` for turning effects on and off based on states
- QuickBuilder, QuickTimeline improvements for creating complex characters with effects and animations
- Better imGui painters
- EditorWorldSystem + editor imGui painters for adding editor UI to your game/app

+++ PhaseJump/PhaseJump/Classes/Render/OpenGL/SomeGLRenderEngine.h
@@ -118,7 +118,7 @@ namespace PJ {
 
         VectorList<SP<GLRenderPlan>> renderPlans;
 
-        ColorFormat colorFormat = ColorFormat::Float;
+        ColorFormat colorFormat = ColorFormat::Byte;
 
         SomeGLRenderEngine();
